+++ modules/security/providers.tf
@@ -1,0 +1,11 @@
+
+terraform {
+  required_version = ">= 1.9"
+  required_providers {
+    # https://registry.terraform.io/providers/hashicorp/aws/latest
+    aws = {
+      source  = "hashicorp/aws"
+      version = "5.94.1"
+    }
+  }
+}