+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next
 
-- Fix: module bucket_private to stay current
+- Fix: Module "bucket_private" to stay current
+- Add: Module "security" with AWS Config, Security Hub, GuardDuty, Inspector, Detective, etc.
 - Add: Cloudflare Security Group for HTTP/HTTPS
 - Edit: Switching default region from us-east-2 to us-east-1
 - Edit: Upgrading RHEL9 AMI from v9.4.x to v9.5.x